fix: refactor ledger checkpoint handling and add period type helpers
