+++ src/v4vapp_backend_v2/accounting/ledger_checkpoints.py
@@ -29,10 +29,8 @@ Each document covers one account × one period:
 from __future__ import annotations
 
 import asyncio
-import calendar
-from datetime import date, datetime, timedelta, timezone
+from datetime import datetime, timedelta, timezone
 from decimal import Decimal
-from enum import StrEnum
 from timeit import default_timer as timer
 from typing import Any, Dict, List, Optional, Tuple
 
@@ -48,14 +46,13 @@ from v4vapp_backend_v2.accounting.account_balances import (
 from v4vapp_backend_v2.accounting.ledger_account_classes import LedgerAccount
 from v4vapp_backend_v2.accounting.ledger_entry_class import LedgerEntry
 from v4vapp_backend_v2.config.setup import InternalConfig, logger
+from v4vapp_backend_v2.helpers.period_end_type import (
+    PeriodType,
+    completed_period_ends_since,
+    last_completed_period_end,
+)
 
 ICON = "📌"
-
-
-class PeriodType(StrEnum):
-    DAILY = "daily"
-    WEEKLY = "weekly"
-    MONTHLY = "monthly"
 
 
 class CheckpointConvSummary(BaseModel):
@@ -252,17 +249,17 @@ async def delete_all_ledger_checkpoints(
 
     """
     try:
-        await LedgerCheckpoint.collection().delete_many(
+        delete_result = await LedgerCheckpoint.collection().delete_many(
             filter={
                 "account_name": account_name,
                 "account_sub": account_sub,
                 "account_type": account_type,
-                "period_type": str(period_type.value),
+                "period_type": str(period_type),
             }
         )
         logger.debug(
             f"📌 Checkpoint deleted: {account_name}:{account_sub} {period_type}",
-            extra={"notification": False},
+            extra={"notification": False, "delete_result": delete_result.raw_result},
         )
     except Exception as e:
         logger.error(
@@ -273,106 +270,57 @@ async def delete_all_ledger_checkpoints(
 
 
 # ---------------------------------------------------------------------------
-# Period boundary helpers
-# ---------------------------------------------------------------------------
-
-
-def period_end_for_date(period_type: PeriodType, d: date) -> datetime:
-    """Return the last microsecond of the calendar period that contains *d*.
-
-    The returned datetime is timezone-aware (UTC) and represents the inclusive
-    upper bound: all transactions with ``timestamp ≤ period_end`` belong to
-    this period's checkpoint.
-    """
-    if period_type == PeriodType.DAILY:
-        return datetime(d.year, d.month, d.day, 23, 59, 59, 999999, tzinfo=timezone.utc)
-
-    if period_type == PeriodType.WEEKLY:
-        # ISO week: Monday=0 … Sunday=6; advance to Sunday
-        days_to_sunday = 6 - d.weekday()
-        sunday = d + timedelta(days=days_to_sunday)
-        return datetime(
-            sunday.year, sunday.month, sunday.day, 23, 59, 59, 999999, tzinfo=timezone.utc
-        )
-
-    if period_type == PeriodType.MONTHLY:
-        last_day = calendar.monthrange(d.year, d.month)[1]
-        return datetime(d.year, d.month, last_day, 23, 59, 59, 999999, tzinfo=timezone.utc)
-
-    raise ValueError(f"Unknown period type: {period_type}")
-
-
-def last_completed_period_end(period_type: PeriodType, now: datetime | None = None) -> datetime:
-    """Return the end datetime of the last fully completed period before *now*.
-
-    Unlike :func:`period_end_for_date`, which returns the end of the period
-    *containing* a given date (which may be in the future), this function
-    always returns a period boundary that has already passed:
-
-    - **daily**  → end of yesterday
-    - **weekly** → end of the most-recently completed ISO week (last Sunday)
-    - **monthly** → end of the previous calendar month
-    """
-    if now is None:
-        now = datetime.now(tz=timezone.utc)
-    today = now.date()
-
-    if period_type == PeriodType.DAILY:
-        d = today - timedelta(days=1)
-    elif period_type == PeriodType.WEEKLY:
-        # weekday(): Mon=0 … Sun=6; go back enough days to land on the last Sunday
-        days_since_last_sunday = today.weekday() + 1  # Mon→1, Tue→2, …, Sun→7
-        d = today - timedelta(days=days_since_last_sunday)
-    elif period_type == PeriodType.MONTHLY:
-        # First day of this month minus one day = last day of previous month
-        d = date(today.year, today.month, 1) - timedelta(days=1)
-    else:
-        raise ValueError(f"Unknown period type: {period_type}")
-
-    return period_end_for_date(period_type, d)
-
-
-def completed_period_ends_since(
-    period_type: PeriodType,
-    since: datetime,
-    until: datetime,
-) -> List[datetime]:
-    """Return a list of completed period-end timestamps in chronological order.
-
-    Only periods whose ``period_end`` is strictly less than *until* are
-    included, so the caller's "current" period is never returned as complete.
-
-    Args:
-        period_type: Granularity of the periods to enumerate.
-        since: Start of the range (inclusive).
-        until: End of the range (exclusive).  Typically *now*.
-    """
-    results: List[datetime] = []
-    current = since.date()
-    while True:
-        end = period_end_for_date(period_type, current)
-        if end >= until:
-            break
-        if end > since:
-            results.append(end)
-
-        # Advance to the next period start
-        if period_type == PeriodType.DAILY:
-            current = current + timedelta(days=1)
-        elif period_type == PeriodType.WEEKLY:
-            days_to_next_monday = 7 - current.weekday()
-            current = current + timedelta(days=days_to_next_monday)
-        elif period_type == PeriodType.MONTHLY:
-            last_day = calendar.monthrange(current.year, current.month)[1]
-            first_next = date(current.year, current.month, last_day) + timedelta(days=1)
-            current = first_next
-
-    return results
-
-
-# ---------------------------------------------------------------------------
 # CRUD helpers
 # ---------------------------------------------------------------------------
+
+
+async def invalidate_checkpoints_for_accounts_by_date(
+    accounts: List[LedgerAccount], timestamp: datetime
+) -> None:
+    """
+    Invalidate the cache for given accounts and timestamp.
+
+    This is necessary after reversing a transaction or making a backdated correction (such as for the
+    Customer Deposits Hive (Asset) account for the server when reversing Limit Order Create transactions).
+
+    Args:
+        - *accounts*: List of LedgerAccount instances identifying the accounts
+        - *timestamp*: The timestamp to compare against the last completed period end
+    """
+    tasks = []
+    # convert timestamp to UTC if it is naive
+    if timestamp.tzinfo is None:
+        timestamp = timestamp.replace(tzinfo=timezone.utc)
+    for period_type in PeriodType:
+        if timestamp < last_completed_period_end(period_type):
+            tasks.append(delete_checkpoints_for_accounts_and_period_type(accounts, period_type))
+    await asyncio.gather(*tasks)
+
+
+async def delete_checkpoints_for_accounts_and_period_type(
+    accounts: List[LedgerAccount], period_type: PeriodType
+) -> None:
+    """
+    Delete all checkpoints for given accounts and period type.
+
+    This is necessary after reversing a transaction or making a backdated correction (such as for the
+    Customer Deposits Hive (Asset) account for the server when reversing Limit Order Create transactions).
+
+    Args:
+        - *accounts*: List of LedgerAccount instances identifying the accounts
+        - *period_type*: Granularity of the checkpoints to delete (daily/weekly/monthly)
+    """
+    tasks = []
+    for account in accounts:
+        tasks.append(
+            delete_all_ledger_checkpoints(
+                account_name=account.name,
+                account_sub=account.sub,
+                account_type=str(account.account_type),
+                period_type=period_type,
+            )
+        )
+    await asyncio.gather(*tasks)
 
 
 async def get_latest_checkpoint_before(
