+++ tests/accounting/test_ledger_checkpoints.py
@@ -24,18 +24,20 @@ from v4vapp_backend_v2.accounting.ledger_account_classes import LiabilityAccount
 from v4vapp_backend_v2.accounting.ledger_checkpoints import (
     CheckpointConvSummary,
     LedgerCheckpoint,
-    PeriodType,
     build_checkpoints_for_period,
-    completed_period_ends_since,
     create_checkpoint,
     delete_all_ledger_checkpoints,
     get_latest_checkpoint_before,
-    last_completed_period_end,
-    period_end_for_date,
 )
 from v4vapp_backend_v2.accounting.ledger_entry_class import LedgerEntry
 from v4vapp_backend_v2.config.setup import InternalConfig
 from v4vapp_backend_v2.database.db_pymongo import DBConn
+from v4vapp_backend_v2.helpers.period_end_type import (
+    PeriodType,
+    completed_period_ends_since,
+    last_completed_period_end,
+    period_end_for_date,
+)
 
 # ---------------------------------------------------------------------------
 # Module-scoped fixtures (DB + test data)
